move operation type to package : operation & add  Subscribe & Renew & Unsubscribe action
